+++ example/datasets/mnist_dataset.h
@@ -9,7 +9,7 @@
 namespace RitsuDataSet {
 
 	extern void loadMNIST(const std::string &imagePath, const std::string &labelPath, const std::string &imageTestPath,
-						  const std::string &labelTestPath, Ritsu::Tensor<float> &dataX, Ritsu::Tensor<uint8_t> &dataY,
-						  Ritsu::Tensor<float> &testX, Ritsu::Tensor<uint8_t> &testY);
+						  const std::string &labelTestPath, Ritsu::Tensor<uint8_t> &dataX, Ritsu::Tensor<uint8_t> &dataY,
+						  Ritsu::Tensor<uint8_t> &testX, Ritsu::Tensor<uint8_t> &testY);
 
 } // namespace RitsuDataSet
